remove unnecessary cast

+++ uvm/impl/com/untangle/uvm/ConntrackMonitorImpl.java
@@ -22,10 +22,10 @@ import com.untangle.uvm.util.Pulse;
 
 public class ConntrackMonitorImpl
 {
-    private static final int CONNTRACK_PULSE_FREQUENCY_MS = 60*1000; /* 1 minute */
-    private static final int CONNTRACK_PULSE_FREQUENCY_SEC = 60; /* 1 minute */
-    private static final int CLEANER_PULSE_FREQUENCY = 560*1000; /* 5 minutes */
-    private static final long LIFETIME_MS = 1000*60*3; /* 2 minutes */ /* Amount of time to keep complete sessions in table */
+    private static final int   CONNTRACK_PULSE_FREQUENCY_MS = 60*1000; /* 1 minute */
+    private static final float CONNTRACK_PULSE_FREQUENCY_SEC = 60f; /* 1 minute */
+    private static final int   CLEANER_PULSE_FREQUENCY = 560*1000; /* 5 minutes */
+    private static final long  LIFETIME_MS = 1000*60*3; /* 2 minutes */ /* Amount of time to keep complete sessions in table */
     private static final Logger logger = Logger.getLogger(ConntrackMonitorImpl.class);
     
     private static ConntrackMonitorImpl INSTANCE = null;
@@ -122,9 +122,9 @@ public class ConntrackMonitorImpl
             long diffC2sBytes = newC2sBytes - oldC2sBytes;
             long diffS2cBytes = newS2cBytes - oldS2cBytes;
             long diffTotalBytes = newTotalBytes - oldTotalBytes;
-            float c2sRateBps = (diffC2sBytes/((float)CONNTRACK_PULSE_FREQUENCY_SEC));
-            float s2cRateBps = (diffS2cBytes/((float)CONNTRACK_PULSE_FREQUENCY_SEC));
-            float totalRateBps = (diffTotalBytes/((float)CONNTRACK_PULSE_FREQUENCY_SEC));
+            float c2sRateBps = (diffC2sBytes/(CONNTRACK_PULSE_FREQUENCY_SEC));
+            float s2cRateBps = (diffS2cBytes/(CONNTRACK_PULSE_FREQUENCY_SEC));
+            float totalRateBps = (diffTotalBytes/(CONNTRACK_PULSE_FREQUENCY_SEC));
 
             /**
              * In some cases specifically UDP a new session takes the place of an old session with the same tuple
